+++ netbox_proxy_plugin/__init__.py
@@ -1,6 +1,6 @@
 from netbox.plugins import PluginConfig
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 
 
 class NetboxProxyPluginConfig(PluginConfig):